requirements.txt
# Core dependencies
streamlit==1.32.0
pandas==2.2.1
Pillow==10.2.0
bcrypt==4.1.2
python-dotenv==1.0.0

# UI Components
streamlit-option-menu==0.3.12
streamlit-lottie==0.0.5
streamlit-image-select==0.6.0
streamlit-cropper==0.1.3

# Analytics
plotly==5.18.0

# Image Processing
opencv-python==4.9.0.80

# Messaging dependencies
slixmpp==1.8.3
aioxmpp==0.13.3
asyncio==3.4.3 